# api/openapi.yaml
openapi: 3.1.0
info:
  title: Rackd API
  description: Infrastructure management API for devices, networks, datacenters, and discovery
  version: 1.0.0
  contact:
    name: Rackd
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 100
        maximum: 1000
    offsetParam:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
...
  schemas:
...
    Datacenter:
      type: object
      required: [id, name, location, description, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        location: { type: string }
        description: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      additionalProperties: true
    DatacenterInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        location: { type: string }
        description: { type: string }
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /api/datacenters:
    get:
      operationId: listDatacenters
      tags: [Datacenters]
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
        - name: name
          in: query
          schema: { type: string }
      responses:
        '200':
          description: List of datacenters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Datacenter'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      operationId: createDatacenter
      tags: [Datacenters]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatacenterInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Datacenter'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/datacenters/{id}:
    parameters:
      - $ref: '#/components/parameters/idPath'
    get:
      operationId: getDatacenter
      tags: [Datacenters]
      responses:
        '200':
          description: Datacenter details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Datacenter'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      operationId: updateDatacenter
      tags: [Datacenters]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatacenterInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Datacenter'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      operationId: deleteDatacenter
      tags: [Datacenters]
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/datacenters/{id}/devices:
    parameters:
      - $ref: '#/components/parameters/idPath'
    get:
      operationId: listDatacenterDevices
      tags: [Datacenters]
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
      responses:
        '200':
          description: Devices in datacenter
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }

  # ── Networks ──
  /api/networks:
    get:
      operationId: listNetworks
      tags: [Networks]
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
        - name: name
          in: query
          schema: { type: string }
        - name: datacenter_id
          in: query
          schema: { type: string, format: uuid }
        - name: vlan_id
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: List of networks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Network'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      operationId: createNetwork
      tags: [Networks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NetworkInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Network'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/networks/{id}:
    parameters:
      - $ref: '#/components/parameters/idPath'
    get:
      operationId: getNetwork
      tags: [Networks]
      responses:
        '200':
          description: Network details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Network'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      operationId: updateNetwork
      tags: [Networks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NetworkInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Network'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      operationId: deleteNetwork
      tags: [Networks]
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/networks/{id}/devices:
    parameters:
      - $ref: '#/components/parameters/idPath'
    get:
      operationId: listNetworkDevices
      tags: [Networks]
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
      responses:
        '200':
          description: Devices in network
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/networks/{id}/utilization:
    parameters:
      - $ref: '#/components/parameters/idPath'
    get:
      operationId: getNetworkUtilization
      tags: [Networks]
      responses:
        '200':
          description: Utilization stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetworkUtilization'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/networks/{id}/pools:
    parameters:
      - $ref: '#/components/parameters/idPath'
    get:
      operationId: listNetworkPools
      tags: [Pools]
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
      responses:
        '200':
          description: Pools in network
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NetworkPool'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      operationId: createNetworkPool
      tags: [Pools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PoolInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetworkPool'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }

  # ── Pools ──
  /api/pools/{id}:
    parameters:
      - $ref: '#/components/parameters/idPath'
    get:
      operationId: getPool
      tags: [Pools]
      responses:
        '200':
          description: Pool details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetworkPool'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      operationId: updatePool
      tags: [Pools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PoolInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetworkPool'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      operationId: deletePool
      tags: [Pools]
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/pools/{id}/next-ip:
    parameters:
      - $ref: '#/components/parameters/idPath'
    get:
      operationId: getPoolNextIP
      tags: [Pools]
      responses:
        '200':
          description: Next available IP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NextIP'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/pools/{id}/heatmap:
    parameters:
      - $ref: '#/components/parameters/idPath'
    get:
      operationId: getPoolHeatmap
      tags: [Pools]
      responses:
        '200':
          description: IP heatmap
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IPHeatmap'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /api/dashboard:
    get:
      operationId: getDashboard
      tags: [Dashboard]
      responses:
        '200':
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSummary'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
...